Fish scripts
Added scripts that track how many fishes we encountered. We now have an integer score for that in Singletons(it's like a global type of a script that every other scripts can access, so for example right now any scripts in the scene can now see which fish we just encountered and according to that they can change the scene however we want).
For now there's a super simple movement script as well which we can easily change for DK's player movement later